Refactor: 구조 명확성을 위해 'header' 디렉토리를 'navbar'로 변경
