--- digital_game_nomad/shared/components/Header/index.ts
@@ -1,2 +1,0 @@
-export { default as HeaderPresenter } from './presenters/Header.presenter';
-export { default as HeaderContainer } from './containers/Header.container';
