+++ digital_game_nomad/shared/components/NavBar/index.ts
@@ -1,0 +1,1 @@
+export { default as Navbar } from './container/NavbarContainer';
